Changed terminology data -> seed. Moved csv loading to model_generator.

+++ model_generator.py
@@ -1,10 +1,12 @@
 #!/bin/python3
+import csv
 import os
 import tomllib
 import re
 import argparse
 import utils
 import shutil
+from functools import partial
 from zarigueya_context import ZarigueyaContext
 from zarigueya_context import load_default_context
 from mako.template import Template
@@ -98,7 +100,11 @@ def context_from_cmd(args: argparse.Namespace) -> ZarigueyaContext:
     if out_path is None:
         out_path = pjoin(input_path, 'output')
 
-    return ZarigueyaContext(models_path, tmplts_path, data_path, out_path, profile_path, use_case_funcs)
+    return ZarigueyaContext(models_path, tmplts_path, out_path, profile_path, data_path, use_case_funcs)
+
+def get_seed_data(seeds_path, model_name):
+    with open(f'{pjoin(seeds_path, model_name)}.csv', newline='') as seeds_file:
+        yield csv.DictReader(seeds_file)
 
 """
 Find templates and process them.
@@ -106,9 +112,10 @@ Find templates and process them.
 """
 def apply_templates(ctx: ZarigueyaContext):
     # The template parameters is a dict containing the current model's
-    # parameters, the global parameters (gbl), and the type conversions (conv).
+    # parameters, its related data (data --if any), the global parameters (gbl),
+    # and the type conversions (conv).
     tmplts_params = {}
-    # Helps deciding what to do with the file
+    # Flag to prevent processing the same file twice if it matches multiple regex
     file_ready = False
     
     for infile_name in os.listdir(ctx.current_inpath):
@@ -116,7 +123,7 @@ def apply_templates(ctx: ZarigueyaContext):
         # infile_path is its full path.
         infile_path = pjoin(ctx.current_inpath, infile_name)
         # Non-template files are just copied
-        if '.tmplt' not in infile_name:
+        if os.path.isfile(infile_path) and '.tmplt' not in infile_name:
             shutil.copy2(
                 infile_path,
                 ctx.current_outpath)
@@ -160,10 +167,11 @@ def apply_templates(ctx: ZarigueyaContext):
                 raise AttributeError(f"{infile_path}: A parent folder has the properties' double-square brackets [[]] notation, which isn't allowed for models file/folders (single square-bracket []).")
 
             for model in ctx.models:
-                mdetails = utils.load_toml(ctx.models_path, model)
-                ctx.current_model = mdetails
+                mdetails = ctx.models[model]
+                ctx.current_model = model
                 tmplt_params = {
                     **mdetails, 
+                    'seed_data': get_seed_data(ctx.seeds_path, model),
                     'gbl': ctx.gbl,
                     'conv': ctx.conversions
                 }
@@ -174,7 +182,8 @@ def apply_templates(ctx: ZarigueyaContext):
         
         if not file_ready:
             tmplt_params = {
-                'models': ctx.models, 
+                'models': ctx.models,
+                'seed_data': partial(get_seed_data, ctx.seeds_path),
                 'gbl': ctx.gbl,
                 'conv': ctx.conversions
             }
